More reorg, move everything to its final location and get rid of the import in settings
