+++ cityhall/api/db/__init__.py
@@ -11,3 +11,145 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from abc import abstractmethod
+from api.db.auth import Auth
+
+
+class DbState(object):
+    Closed = 0
+    Open = 1
+    Error = 2
+
+
+class Rights(object):
+    DontExist = -1
+    NoRights = 0
+    Read = 1
+    ReadProtected = 2
+    Write = 3
+    Grant = 4
+
+
+class DbFactory(object):
+    """
+    This class is the factory which will create light-weight Db classes which
+    will be used to actually do the work of City Hall.
+    """
+    @abstractmethod
+    def open(self):
+        pass
+
+    @abstractmethod
+    def is_open(self):
+        pass
+
+    @abstractmethod
+    def get_db(self):
+        pass
+
+    @abstractmethod
+    def create_default_tables(self):
+        pass
+
+    @abstractmethod
+    def authenticate(self, user, passhash):
+        pass
+
+
+class Db(object):
+    """
+    The class in this directory wil be expecting to pass around a object that
+    implements these methods.  The idea is to take these and implement them
+    for any database.
+
+    To get a good idea of what the database should be doing behind the scenes,
+    see lib/db/memory/cityhall_db.py and, for a slightly higher-level overview,
+    see its corresponding tests: /test/tests_connection.py
+    """
+    @abstractmethod
+    def create_root(self, author, env):
+        pass
+
+    @abstractmethod
+    def get_children_of(self, index):
+        pass
+
+    @abstractmethod
+    def update(self, user, index, value):
+        pass
+
+    @abstractmethod
+    def create(self, user, parent, name, value, override=''):
+        pass
+
+    @abstractmethod
+    def get_value(self, index):
+        pass
+
+    @abstractmethod
+    def get_value_for(self, parent_index, name, override):
+        pass
+
+    @abstractmethod
+    def get_env_root(self, env):
+        pass
+
+    @abstractmethod
+    def get_history(self, index):
+        pass
+
+    @abstractmethod
+    def delete(self, author, index):
+        pass
+
+    @abstractmethod
+    def set_protect_status(self, author, index, status):
+        pass
+
+    @abstractmethod
+    def create_user(self, author, user, passhash, user_root):
+        pass
+
+    @abstractmethod
+    def update_user(self, author, user, passhash):
+        pass
+
+    @abstractmethod
+    def delete_user(self, author, user):
+        pass
+
+    @abstractmethod
+    def get_users(self, env):
+        pass
+
+    @abstractmethod
+    def get_child(self, parent, name, override=''):
+        pass
+
+
+class Connection(object):
+    def __init__(self, db):
+        self.db_connection = db
+
+    def connect(self):
+        self.db_connection.open()
+
+    def _ensure_open(self):
+        return self.db_connection.is_open()
+
+    def authenticate(self, user, passhash):
+        if self._ensure_open():
+            return self.db_connection.authenticate(user, passhash)
+        return None
+
+    def create_default_env(self):
+        if self._ensure_open():
+            self.db_connection.create_default_tables()
+
+    def get_auth(self, user, passhash):
+        open = self._ensure_open()
+        authenticated = self.db_connection.authenticate(user, passhash)
+        if open and authenticated:
+            return Auth(self.db_connection.get_db(), user, authenticated)
+        return None
