+++ cityhall/api/db/django/users.py
@@ -14,27 +14,12 @@
 
 from copy import copy
 from django.db import transaction
-from api.models import User
-from django.db import models
-
-
-class EnvUsers(models.Model):
-    name = models.TextField(max_length=64)
-    entry = models.TextField(max_length=2048)
+from api.models import User, EnvUsers
 
 
 class Users(object):
     def get_users(self, env):
-        ret = EnvUsers.objects.raw(
-            'SELECT u.id, u.name, v.entry '
-            'FROM api_user u, api_value v '
-            'WHERE u.user_root = v.parent '
-            '      and u.active = %s '
-            '      and v.active = %s '
-            '      and v.name = %s ',
-            [True, True, env]
-        )
-
+        ret = EnvUsers.get_users(env)
         return {u.name: u.entry for u in ret}
 
     def create_user(self, author, user, passhash, user_root):
