+++ cityhall/test/test_auth.py
@@ -13,9 +13,8 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from django.test import TestCase
+from api.db import Connection, Rights
 from api.db.memory.db_factory import CityHallDbFactory
-from api.db.connection import Connection
-from api.db.constants import Rights
 
 
 class TestAuthentication(TestCase):
