+++ cityhall/test/test_connection.py
@@ -13,8 +13,8 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from django.test import TestCase
+from api.db import Connection
 from api.db.memory.db_factory import CityHallDbFactory
-from api.db.connection import Connection
 
 
 class TestConnection(TestCase):
