+++ cityhall/test/test_env.py
@@ -14,8 +14,7 @@
 
 from django.test import TestCase
 from api.db.memory.db_factory import CityHallDbFactory
-from api.db.constants import Rights
-from api.db.connection import Connection
+from api.db import Rights, Connection
 
 
 class TestEnvironment(TestCase):
