+++ cityhall/test/test_memory_db.py
@@ -13,7 +13,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from django.test import TestCase
-from api.db.constants import DbState, Rights
+from api.db import DbState, Rights
 from api.db.memory.db_factory import CityHallDbFactory
